Move Asset abc.py -> asset.py
