+++ discord/__init__.py
@@ -29,6 +29,7 @@ from .guild import Guild
 from .relationship import Relationship
 from .member import Member, VoiceState
 from .message import Message, Attachment
+from .asset import Asset
 from .errors import *
 from .calls import CallMessage, GroupCall
 from .permissions import Permissions, PermissionOverwrite
