+++ docs/api.rst
@@ -1855,8 +1855,6 @@ module, others which are not.
 
 .. autoclass:: discord.abc.Connectable
 
-.. autoclass:: discord.abc.Asset
-
 .. _discord_api_models:
 
 Discord Models
@@ -1918,6 +1916,12 @@ Attachment
 ~~~~~~~~~~~
 
 .. autoclass:: Attachment()
+    :members:
+
+Asset
+~~~~~
+
+.. autoclass:: Asset()
     :members:
 
 Message
